Add files via upload
This project is based on countries dataset, which is been extracted from Kaggle website. This project is to cluster the countries who is fighting against the poverty and to obtain the overall development of their countries. They are in need of help from HELP International (international humanitarian NGO). I have used unsupervised learning algorithms to form the clusters.